refactoring and splitting code for maintainability
